refactor(settings): move font + raw-markdown controls to Editor tab
Editor Font, Font by File Type, and Show Raw Markdown govern how the
document reads, not how the app looks, so per #1262's decided scope
they move out of Appearance (which had become a dumping ground) and
into Editor. Show Formatting Bar stays in Appearance per the same
decision. testid appearance-show-raw-markdown keeps its historical
name — Critical Rule 7 forbids renaming a data-testid.
