+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - **New setups no longer register the channel helper unless you ask for it.** Until now, running Tandem's setup wrote a second connection — the "channel shim" — into your Claude configuration whether or not you were in a position to use it. That helper only delivers if you also start Claude Code with the `--dangerously-load-development-channels` flag, and most people don't, so for most installs it connected, sat there, and delivered nothing. It was not merely idle: Tandem decides whether to warn you that *nothing* is listening by checking whether anything is attached, and a helper that is attached-but-deaf answers that question wrongly. The warning built to catch exactly this situation was therefore silenced by the situation itself. Setup now writes it only when you ask — `tandem setup --apply --with-channel-shim` — and the two mechanisms that actually deliver are untouched: sessions Tandem starts for you are woken directly, and sessions you start yourself can watch for updates on their own. **If you already have it configured, nothing changes and nothing is removed** — including when you rotate your connection token, which briefly would have deleted it. Tandem cannot tell an entry it wrote by default from one you deliberately asked for — the two are identical on disk — so rather than guess, it leaves yours alone. Re-run `tandem setup --apply` without the flag if you want it gone.
 
+- **Editor Font, Font by File Type, and Show Raw Markdown moved from Settings → Appearance to Settings → Editor (#1262).** Appearance was becoming a dumping ground for anything that touched text, while Editor sat nearly empty. These three control how the document itself reads, not how the app chrome looks, so they now live where you'd look for them. Theme, Text Size, Accent Color, Spacing Density, Decorations, formatting bar and tab width are unaffected.
+
 - **We stopped saying the flag is temporary.** Older text implied it would become unnecessary once Claude Code's real-time API settled. It won't: the list that would make it optional covers plugins only, and Tandem's channel is a plain connection rather than a plugin, so there is nothing for it to be added to. Better to say so than to leave people waiting.
 
 ### Security
